# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/internal/v1/user": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "新增",
                "parameters": [
                    {
                        "description": "會員資料",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/internal/v1/user/{id}": {
            "delete": {
                "description": "說明",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "刪除",
                "responses": {}
            },
            "patch": {
                "description": "說明",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "部份更新",
                "responses": {}
            }
        },
        "/private/v1/user/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "查單筆",
                "parameters": [
                    {
                        "type": "string",
                        "description": "會員id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/v1/users": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用戶"
                ],
                "summary": "查多筆",
                "parameters": [
                    {
                        "type": "string",
                        "description": "會員姓名",
                        "name": "name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "model.User": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "識別碼",
                    "type": "string",
                    "example": "123"
                },
                "name": {
                    "description": "姓名",
                    "type": "string",
                    "example": "david"
                }
            }
        }
    }
}